+++ src/main/java/hello/roommate/member/controller/MemberController.java
@@ -274,41 +274,4 @@ public class MemberController {
 		result.sort(((o1, o2) -> o2.getUpdatedTime().compareTo(o1.getUpdatedTime())));
 		return result;
 	}
-
-	/**
-	 * 기본 추천 목록 반환
-	 *
-	 * @param memberId 사용자 id
-	 * @return 추천 목록 멤버 반환
-	 */
-	@GetMapping("/{memberId}/recommendation")
-	public List<RecommendMemberDTO> recommendMembers(@PathVariable Long memberId) {
-		log.info("추천 목록 반환 요청, id={}", memberId);
-		List<Member> members = memberService.recommendMembers(memberId);
-
-		//dto로 변환
-		List<RecommendMemberDTO> dtoList = members.stream()
-			.map(member -> memberService.convertToDTO(member))
-			.collect(Collectors.toList());
-		log.info("{}", dtoList);
-		return dtoList;
-	}
-
-	/**
-	 * 필터 적용하여 추천목록 반환
-	 *
-	 * @param memberId 사용자 id
-	 * @param filterCond 사용자가 적용한 필터 항목들
-	 * @return 필터 적용된 추천 목록 멤버 반환
-	 */
-	@PostMapping("/{memberId}/recommendation")
-	public List<RecommendMemberDTO> searchMembers(@PathVariable Long memberId,
-		@RequestBody @Validated FilterCond filterCond) {
-		List<Member> members = memberService.searchMembers(memberId, filterCond);
-		List<RecommendMemberDTO> dtoList = members.stream()
-			.map(member -> memberService.convertToDTO(member))
-			.collect(Collectors.toList());
-
-		return dtoList;
-	}
 }
