+++ src/main/java/hello/roommate/member/repository/MemberCustomRepository.java
@@ -8,4 +8,5 @@ import hello.roommate.recommendation.domain.enums.Category;
 
 public interface MemberCustomRepository {
 	List<Member> search(Long memberId, Map<Category, List<Long>> cond, List<Integer> ages);
+
 }
